% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auto_Importer.R
\name{Auto_Importer}
\alias{Auto_Importer}
\title{A Function to Automatically Import Files in a Folder}
\usage{
Auto_Importer(
  directory,
  file_type = ".csv",
  Encoding = "UTF-8",
  Deliminator = ",",
  Keyword = FALSE,
  Exclude = FALSE,
  inList = FALSE,
  Meta_List_Name = "df_list",
  ...
)
}
\arguments{
\item{directory}{Directory of where the files that you want to import <e.g. "C:/Users/___YOUR USERNAME___/UPSTREAM FOLDER/.../FOLDER NAME/">}

\item{file_type}{The type of file to import. Currently, supports  ".tsv  ".csv", ".xls(x)",".sas",".sav". <default: ".csv">}

\item{Encoding}{If .csv files would be imported, one can specify}

\item{Deliminator}{If the data is stored in .csv file and the Deliminator is not comma as default, please, specify it. <default: ",">}

\item{Keyword}{Specify keywords to import only specific files based on file names <Default: FALSE, i.e., import all>}

\item{Exclude}{Specify keywords to exclude specific files from importation <Default: FALSE, i.e., exclude none>}

\item{inList}{Binary operator identify if to import all files into a large list <Default: FALSE, i.e., import each separately into individual dataframe>}

\item{Meta_List_Name}{The name assigned to the meta data list if inList is TRUE <Default: df_list>}
}
\value{
Import files into your working enviroment in R (i.e., Global Enviroment in R Studio)
}
\description{
This function will loop over logistic regressions in a data frame specified with at least 1 unadjusted model and an adjusted one, cross-sectionally.
To avoid phaking, the pipeline will automatically detect underpower study and aborpt. Please, check Peduzzi 1996 for calculating statistical powers.
}
\examples{
Download the folder from
Auto_Importer(Directory_to_Downloads_Folder)


}
