% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Butler.R
\name{Butler}
\alias{Butler}
\title{Butler}
\usage{
Butler(Task, Keep = NULL)
}
\arguments{
\item{Task}{A logical operator to assign one of the pre-specified tasks to the butler (this function). The tasks include
\enumerate{
\item "Book", to document all the items in the current global environment
\item "Clean", to clean up everything not documented in the butler's book (a.k.a., BBook). The Butler will also update the BBook by removing manually deleted objects previously documented.
\item "Clear", to clear the output and the console.
}}

\item{Keep}{A place-holding text vector, that allows users to specify the new/undocumented items to keep while cleaning. Besides the names of the items, users may also passed class categories to restrict the cleaning range. For instance, when "table" is passed, the Butler will keep everything previous documented and any new table-class item.}
}
\description{
A function to help keeping the global environment organized
}
\examples{

# First, show the variables listed in the global environment
ListAll = ls()

## If the Global Environment is clean, then create some things
if(length(ListAll)<1){

  data("SimSleep") # A data.list

  WD = getwd() # A Directory

  Date = Sys.Date() # A Variable

  Hello_World = function(){ print("Hello_World") } # A Function
}

# Generate a Documentation, called "BBook", which stands for Butler's Book.
Butler(Task="Book")

# Add new items to the environment
data("iris")

AVector = stats::rnorm(100)

# Clean-up and Go to the Documented Environment
Butler(Task="Clean") # To keep certain "new" items, specify the name(s) of
# the object(s) or item class(es) via Keep.



}
\keyword{clean}
\keyword{document}
