% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clear_History.R
\name{Clear_History}
\alias{Clear_History}
\title{A Function to Document Executed Codes by Date with Continuous Time Stamps Separating Different Entries}
\usage{
Clear_History(Directory)
}
\arguments{
\item{Directory}{Directory to where the executed codes will be documented. <e.g. "C:/Users/___YOUR USERNAME___/UPSTREAM FOLDER/.../FOLDER NAME/">}
}
\value{
Current time stamp
}
\description{
This function was designed to facilitate lab note documentation while help maintain a minimum clean history window. Executed codes will be stored automatically in the History folder in the directory assigned.
}
\examples{
Clear_History(getwd())


}
