% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Col_Name_Checker.R
\name{Col_Name_Checker}
\alias{Col_Name_Checker}
\title{An Automatic Pipeline to Check and Correct Column Names}
\usage{
Col_Name_Checker(col.names, ...)
}
\arguments{
\item{col.names}{The name of variables/columns within a dataframe.}
}
\value{
'Corrected' Names for Dataframes starting with non-alphabatic characters
}
\description{
This function will loop over all data frames in the work environment. If a dataframe starts with a digit or a punctuation, the pipeline will add a "x" in
front of it so the user will not need to wrap `` around the name of a dataframe during computation.
}
