% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDA.R
\name{create_mi}
\alias{create_mi}
\title{Create missing indicator for cols in a data.table}
\usage{
create_mi(dt, cols = colnames(dt), suffix = "_mi", detect = c(F, T))
}
\arguments{
\item{dt}{a data.table}

\item{cols}{column names to create missing indicator for, default to all columns}

\item{suffix}{the suffix to append to the column name to form the name for missing indicator columns}

\item{detect}{Logical value specifying if missing indicator is created for all columns with missing values}
}
\value{
None. All data manipulation is done in place.
}
\description{
Create missing indicator for cols in a data.table
}
