% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DF_Name_Checker.R, R/Name_Checker.R
\name{DF_Name_Checker}
\alias{DF_Name_Checker}
\title{An Automatic Pipeline to Check and Correct Dataframe Names}
\usage{
DF_Name_Checker(Simple = TRUE, ...)

DF_Name_Checker(Simple = TRUE, ...)
}
\arguments{
\item{Simple}{An operator determines whether to simply add a cap at the begining. If false, all texts before the first alphaba will be moved to the end. <default: TRUE>}
}
\value{
'Corrected' Names for Dataframes starting with non-alphabatic characters

'Corrected' Names for Dataframes starting with non-alphabatic characters
}
\description{
This function will loop over all data frames in the work environment. If a dataframe starts with a digit or a punctuation, the pipeline will add a "x" in
front of it so the user will not need to wrap `` around the name of a dataframe during computation.

This function will loop over all data frames in the work environment. If a dataframe starts with a digit or a punctuation, the pipeline will add a "x" in
front of it so the user will not need to wrap `` around the name of a dataframe during computation.
}
