% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirChecker.R
\name{DirChecker}
\alias{DirChecker}
\title{Check if the directory is a file or a folder}
\usage{
DirChecker(Dir)
}
\arguments{
\item{Dir}{Directory of where the files that you want to import (e.g. "C:/Users/\emph{\strong{YOUR USERNAME}}/UPSTREAM FOLDER/.../FOLDER NAME/"). If not specified, the default is set to the working directory.}
}
\value{
Directory class
}
\description{
Check if the directory is a file or a folder
}
