% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDA.R
\name{dTail}
\alias{dTail}
\title{Computes stats useful for right-skewed, heavy-tailed, zero-inflated variables}
\usage{
dTail(cols, DT)
}
\arguments{
\item{cols}{a character vector of column names}

\item{DT}{data.table object containing cols}
}
\value{
a data.table object containing the summary stats
}
\description{
Computes stats useful for right-skewed, heavy-tailed, zero-inflated variables
}
