% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Historian.R
\name{Historian}
\alias{Historian}
\title{Historian}
\usage{
Historian(Directory)
}
\arguments{
\item{Directory}{Directory to where the executed codes will be documented. <e.g. "C:/Users/\emph{\strong{YOUR USERNAME}}/UPSTREAM FOLDER/.../FOLDER NAME/">}
}
\value{
Current time stamp
}
\description{
A Function to Document Executed Codes by Date with Continuous Time Stamps Separating Different Entries
}
\details{
This function was designed to facilitate lab note documentation while help maintain a minimum clean history window. Executed codes will be stored automatically in the History folder in the directory assigned.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

# Find working directory
getwd()

# Add a documentation...
Historian()
\dontshow{\}) # examplesIf}
}
