% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Importer.R
\name{Importer}
\alias{Importer}
\title{A Function to Automatically Import Files in a Folder}
\usage{
Importer(
  Directory = NULL,
  File_Type = "csv",
  Encoding = "",
  Delim = ",",
  Keyword = NULL,
  Exclude = NULL,
  inList = FALSE,
  Meta_List_Name = "df_list",
  Progress = FALSE
)
}
\arguments{
\item{Directory}{Directory of where the files that you want to import (e.g. "C:/Users/\emph{\strong{YOUR USERNAME}}/UPSTREAM FOLDER/.../FOLDER NAME/"). If not specified, the default is set to the working directory.}

\item{File_Type}{The type of file to import. Currently, supports  "txt", "tsv", "csv", "xls(x)","sas", and "sav". <default: "csv">}

\item{Encoding}{If csv files would be imported, one can specify}

\item{Delim}{If the data is stored in csv file and the Delim is not comma as default, please, specify it. Common deliminators include comma "," semicolon ";" colon ":" tab "/t" .  <default: ",">}

\item{Keyword}{Specify keywords to import only specific files based on file names <Default: FALSE, i.e., import all>}

\item{Exclude}{Specify keywords to exclude specific files from importation <Default: FALSE, i.e., exclude none>}

\item{inList}{Binary operator identify if to import all files into a large list <Default: FALSE, i.e., import each separately into individual dataframe>}

\item{Meta_List_Name}{The name assigned to the meta data list if inList is TRUE <Default: df_list>}
}
\value{
Import files into your working enviroment in R (i.e., Global Enviroment in R Studio)
}
\description{
This function imports files with the same file type as specified in bulk. Users are required to provide the Directory to the folder containing files to be imported.
}
\details{
Importer supports most of the commonly used data format, including  "txt", "tsv", "csv", "xls(x)","sas", and "sav".

All sheets contained within an excel file will be imported into a data list structure while keeping the assigned sheet names for each sub-list.

Users can also specify if the imported files should be merged into a data list structure or as separate data.frame.
If unsure, users may be able to determine by simply opening one of the files in notepad++. Common Delims include tabs "/t" , comma "," , semicolon ";" , and period "." .
}
\examples{

# Load the Simulated Sleep Datasets
data("SimSleep")

# Change the Working Directory to the Default Temporary Directory
Current.Dir <- getwd(); setwd(tempdir())

# Export ALL 20 Simulated Data to Individual Files
for(i in names(SimSleep)){
write.table(SimSleep[[i]], file = paste0(i,".txt"), sep = "," )
# sep allows users to specify the desired deliminater.
}

# Import All 20 Files at the Same Time
# < Option 1 > One data.frame per file (i.e., default option)
Importer(Directory = getwd(),
              File_Type = "txt",
              Delim = ",")

# < Option 2 > Store All 20 files into a data.list
Importer(Directory = getwd(),
              File_Type = "txt",
              Delim = ",",
              inList = TRUE,
              Meta_List_Name = "SimSleep2")

# Reset Working Directory
setwd(Current.Dir)
rm(Current.Dir)

}
\seealso{
\code{\link{Merger}}
}
\keyword{import,}
\keyword{read}
