% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inject.R
\name{Inject}
\alias{Inject}
\title{Inject}
\usage{
Inject(x, Index, Value)
}
\arguments{
\item{x}{A vector of numeric/integer values}

\item{Index}{A vector of index for either row ID or sequential ID of a vector}

\item{Value}{A value or a vector of values to be added.}
}
\value{
A new vector with values injected
}
\description{
A function to inject specified values based on the index
}
\details{
By providing the index (e.g., sequential ID of a vector or rowID of a
data.frame/table), the function will add the specified values between i
indice and the indice before it.
}
\examples{

# An example to inject NA into the vector x
x = sample(c(1:10), 1000, replace = TRUE)
ID = sample(1:1000, 10)

x1 = Inject(x, Index = ID, Value = NA)

# An example to inject specific values into the vector x
x2 = Inject(x, Index = ID, Value = sample(100:1000, 10))


print(x1[ID + 1:length(ID)])
print(x2[ID + 1:length(ID)])

}
\keyword{add,}
\keyword{inject}
