% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inject.Row.R
\name{Inject.Row}
\alias{Inject.Row}
\title{A dynamic function to allow users to add rows to specific location within a data.frame with or without values.}
\usage{
Inject.Row(df, Index, Variable = NULL, Value = NULL)
}
\arguments{
\item{df}{A data.frame}

\item{Index}{A vector of index for either row ID or sequential ID of a vector}

\item{Variable}{A variable or a vector of variable names where to values are to be modified.}

\item{Value}{A value or a vector of values to be added.}
}
\value{
A new vector with values injected
}
\description{
A dynamic function to allow users to add rows to specific location within a data.frame with or without values.
}
\examples{

# To inject empty rows into the sleep data
Test = Inject.Row(sleep, sample(1:20,5))

# To inject specific value at specific variable into the sleep data
Test1 = Inject.Row(sleep, sample(1:20,5), Variable = "group", Value = 1)

# To dynamically inject values at variable
Test2 = Inject.Row(sleep, sample(1:20,3), Variable = names(sleep),
                   Value = c("a","b","c"))



}
\keyword{add,}
\keyword{inject}
