% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lexicographer.R
\name{Lexicographer}
\alias{Lexicographer}
\title{Lexicographer is a function to help compose a project-oriented dictionary.}
\usage{
Lexicographer(Directory, Data = NULL, Missing_Identifier = FALSE, ...)
}
\arguments{
\item{Directories}{Directory of where the files that you want to document <e.g. "C:/Users/___YOUR USERNAME___/UPSTREAM FOLDER/.../FOLDER NAME/">}
}
\value{
Import files into your working enviroment in R (i.e., Global Enviroment in R Studio)
}
\description{
This function will automatically document all the variables embeded in the project.
When calling Lexicographer for the first time, Lexicographer will create a dictionary template at where the
When activate Lexicographer, the
}
\examples{
Download the folder from
Clerk(Directory_to_Downloads_Folder)


}
\keyword{Import,}
\keyword{Read}
