% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListTree.R
\name{ListTree}
\alias{ListTree}
\title{A function to print out the names of the sublist and their items}
\usage{
ListTree(object, ...)
}
\arguments{
\item{object}{A list.}

\item{...}{Internal parameter space holder}
}
\description{
This function will loop over all the sublists and their items within a list and print out their names and structures.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

# Create a Test List

List = list(
sublist1 = list(a = 1, b = 2),
sublist2 = list(c = 3, d = list(e = 5, f = 6)),
sublist3 = list(g = 7, h = list(i = 8, j = list(k = 9)))
)


# list all names
ListTree(List)


\dontshow{\}) # examplesIf}
}
