% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Name_Checker.R
\name{Name_Checker}
\alias{Name_Checker}
\title{A Function to Check for Potentially Problematic Variable Names and Correct Them}
\usage{
Name_Checker(Names, Silent = TRUE)
}
\arguments{
\item{Names}{The name of variables/columns within a dataframe.}

\item{Silent}{A logical indicator for reporting summary.}
}
\value{
'Corrected' variable names
}
\description{
This function will loop over all data frames in the work environment. If a dataframe starts with a digit or a
punctuation, the pipeline will add a "x" in front of it so the user will not need to wrap `` around
the name of a data.frame during computation.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

# Create a dummy data.frame
df = data.frame(x = rnorm(100))

df$`Hours in Bed` = abs(rnorm(100)*10)
df$`10 Hrs of Sleep ` = factor(ifelse(df$`Hours in Bed`>10,"Yes","No"))
df$`Subject who Forget to Sign Agreement!` = factor(rbinom(n = 100,size = 1,prob = 0.1))

Name_Checker(names(df))


# The same principle can be applied to correct data.list documentation.


\dontshow{\}) # examplesIf}
}
