% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Naming_Variables.R
\name{Naming_Variables}
\alias{Naming_Variables}
\title{A function listing out variables to faciliate writing codes or functions}
\usage{
Naming_Variables(x, Delim = ",")
}
\arguments{
\item{x}{Values to be passed through and printed out in a string with quotes.}

\item{Delim}{Deliminators to be used to separate individual items within the value x.}
}
\value{
A long string with characters in quotation marks
}
\description{
This is a simple function that will print out values (e.g., strings, names or numbers) while keeping the quotation marks, which were otherwise not retained via paste or cat. This is particularly useful when manually organizing data files or passing co-variates while performing regression models. It can also be helpful when moving large numbers of files via Bash shell (Linux)
}
\examples{
Naming_Variables(names(iris))  #"Sepal.Length","Sepal.Width","Petal.Length","Petal.Width","Species"

Naming_Variables(ls()) # List everything stored in the global environment (i.e., within the "Environment" tab)

Naming_Variables(list.files(getwd())) # List all the files within the current working directory


}
