% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReOrder.R
\name{ReOrder}
\alias{ReOrder}
\title{A function to reorder data.frame by column or row}
\usage{
ReOrder(Data, Order)
}
\arguments{
\item{Data}{a data.frame}

\item{Order}{To reorder data.frame by column, provide a vector of column names in the desired sequential order. To reorder data.frame by row, provide the name of the variable/column to be used to reorder the data.frame}
}
\value{
a reordered data.frame
}
\description{
Based on the pre-specified order of column names or the levels of a factorial variable, ReOrder will reorganize the data.frame.
}
\keyword{order}
\keyword{reorder}
