% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RowBind.R
\name{RowBind}
\alias{RowBind}
\title{RowBind}
\usage{
RowBind(DFx, DFy, Keep = "SAME")
}
\arguments{
\item{DFx}{First data.frame}

\item{DFy}{Second data.frame}

\item{Keep}{A logical indicator for the rwo binding condition. There are Five indicators pre-programmed: "SAME" (keeping only the shared variables), "x" (keeping only the variables existed in the first data.frame), "y" (keeping only the variables existed in the second data.frame), "either" (keeping only the variables existed in one but not both data.frame) and "ALL" (keeping all available varialbes found in either data.frame)}
}
\value{
a new row bond data.frame
}
\description{
A Function to bind data.frame by row
}
\details{
This function is an extension rbind (i.e., row bind or merge) without the traditional limits of column names and their orders. Users can choose to either keep only the columns with the same names ("SAME"), only based on one of the two data.frames ("x", "y", "either") or all existing columns ("ALL").
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

# Create Sample Data
data("sleep")

# Extract variable names
Variables = names(sleep)

# Split the data by half based on the extra hours of sleep
sleep.1 = subset.data.frame(sleep,subset = sleep$extra<2)
# less than 2-hour increase on total sleep time
sleep.2 = subset.data.frame(sleep,subset = sleep$extra>=2)
# more than 2-hour increase on total sleep time

# Generate a variable in sleep.1 dataset to identify negative changes in
# sleep time
sleep.1$SleepReduction  = factor(ifelse(sleep.1$extra < 0, 1, 0))

# First, try using the base code rbind (remove the # sign before executing
# the following line.)
#rbind(sleep.1,sleep.2)

# Now, let's try RowBind
RowBind(sleep.1,sleep.2)
\dontshow{\}) # examplesIf}
}
\keyword{RowBind}
\keyword{bind,}
\keyword{rbind,}
\keyword{row,}
