% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SandBox.R
\name{SandBox}
\alias{SandBox}
\title{SandBox}
\usage{
SandBox(Expr, Show = TRUE, Commit = FALSE)
}
\arguments{
\item{Expr}{Any give length of codes to be executed. It is crucial to wrap {} around the code or code chunks.}

\item{Show}{A logical parameter to allow printing the results of each computation. <Default  = FALSE> Please, note that it is not recommended to use this parameter when computing/processing large matrices or data. The heavy demands to print the results may exhaust the memory.}

\item{Commit}{A logical indicator to specify whether to add the changes back to the Global Envrionment. This is particularly useful when data manipulations procedures (e.g., categorizing a continuous data)}
}
\value{
Nothing unless committed
}
\description{
A Function to create a SandBox (i.e., a temporary environment) to perform code testing or data manipulation without affecting the Global environment
}
\details{
There are two main methods to use the SandBox.
For simple tasks (e.g., categorizing a variable or test a simple function), the function SandBox would be enough.
When performing more complex tasks, please, use PlayGround. Similar to the SandBox, the PlayGround also allows users
to commit changes to the Global Environment. When complete, use Close() to dissolve the PlayGround and save the memory.
}
\examples{

# Create a variable called x
x = 2


# Use the SandBox function to modify x in the sandbox environment and Commit the changes

SandBox({
        x <- x + 1
        print(x)
       })

print(x) # The original x value


SandBox({
        x <- x + 1
       },
       Commit = TRUE,
       Show = TRUE)

}
\keyword{Import,}
\keyword{Read}
