% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update}
\alias{update}
\title{A function to check the latest release version of the package and update it.}
\usage{
update()
}
\description{
This function will check the latest version of RHelper and update if needed.
}
\keyword{RHelper}
\keyword{Update,}
