% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Who_is.R
\name{Who_is}
\alias{Who_is}
\title{A quick function to print the names of all elements within a specific type in the environment.}
\usage{
Who_is(
  Type = c("Data.frame", "Function", "List", "Value"),
  print_out = FALSE,
  ...
)
}
\arguments{
\item{Type}{a type of elements in your environment. <Options: "Data.frame", "Functions", or "List">}

\item{print_out}{to identify whether to print out the names of all elements within the environment. <default: FALSE, which will not be seen on your screen>.}

\item{DA}{0 = do not combine results from MSTP; 1 = combine results from MSTP.}
}
\value{
A savable report with AIC index
}
\description{
This function will print out the names of all data.frame, list, or functions in the environment.
It is particularly useful when wanting to save computer memory to optimize the use of R.
}
\examples{
#Who_is("Data.frame") # List all data.frames in the global environment.

}
