% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auto_Importer.R
\name{Auto_Importer}
\alias{Auto_Importer}
\title{A Function to Automatically Import Files in a Folder}
\usage{
Auto_Importer(
  Directory,
  File_Type = ".csv",
  Encoding = "UTF-8",
  Delim = ",",
  Keyword = FALSE,
  Exclude = FALSE,
  inList = FALSE,
  Meta_List_Name = "df_list",
  Progress = FALSE,
  ...
)
}
\arguments{
\item{Directory}{Directory of where the files that you want to import <e.g. "C:/Users/___YOUR USERNAME___/UPSTREAM FOLDER/.../FOLDER NAME/">}

\item{File_Type}{The type of file to import. Currently, supports  ".txt .tsv  ".csv", ".xls(x)",".sas",".sav". <default: ".csv">}

\item{Encoding}{If .csv files would be imported, one can specify}

\item{Delim}{If the data is stored in .csv file and the Delim is not comma as default, please, specify it. Common deliminators include comma "," semicolon ";" colon ":" tab "/t" .  <default: ",">}

\item{Keyword}{Specify keywords to import only specific files based on file names <Default: FALSE, i.e., import all>}

\item{Exclude}{Specify keywords to exclude specific files from importation <Default: FALSE, i.e., exclude none>}

\item{inList}{Binary operator identify if to import all files into a large list <Default: FALSE, i.e., import each separately into individual dataframe>}

\item{Meta_List_Name}{The name assigned to the meta data list if inList is TRUE <Default: df_list>}
}
\value{
Import files into your working enviroment in R (i.e., Global Enviroment in R Studio)
}
\description{
This function imports files with the same file type as specified in bulk. Users are required to provide the Directory to the folder containing files to be imported. #' Auto_Importer supports most of the commonly used data format, including .txt .tsv .csv, .xls(x), .sas, .sav . All sheets contained within an excel file will be imported into a data list structure while keeping the assigned sheet names for each sub-list. Users can determine if the imported files should be merged into a data list structure or as separate data.frame.
If unsure, users may be able to determine by simply opening one of the files in notepad++. Common Delims include tabs "/t" , comma "," , semicolon ";" , and period "." .
}
\examples{
Download the folder from
Auto_Importer(Directory_to_Downloads_Folder)


}
\keyword{Import,}
\keyword{Read}
