% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Librarian.R
\name{Librarian}
\alias{Librarian}
\title{A function to help prepare a dataset for analysis}
\usage{
Librarian(
  Data,
  Dictionary = NULL,
  Missing_Identifier = NA,
  KeyExclude = NULL,
  NewName = NULL,
  ...
)
}
\arguments{
\item{Data}{a data.frame}

\item{Missing_Identifier}{Values identified as missing}

\item{KeyExclude}{Values to be excluded}

\item{NewName}{Name name for the modified data.}

\item{Directory}{Directory for the project}
}
\value{
Import files into your working enviroment in R (i.e., Global Enviroment in R Studio)
}
\description{
Based on the dictionary prepared by Lexicographer, Librarian will prepare the variables within a dataset ready for analysis.
}
\keyword{librarian}
