% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NCSM_205_1994_Modified_K}
\alias{NCSM_205_1994_Modified_K}
\title{Daily Forecast}
\format{
A simulated data with 100 rows and 5 variables:
\describe{
\item{\code{Age}}{Age at the recuitment}
\item{\code{Sex}}{Participant's biological sex}
\item{\code{K10}}{Psychological distress scale}
\item{\code{Insomnia}}{Insomnia symptom status}
\item{\code{Apnea}}{Apnea symptom status}
}
}
\usage{
data(NCSM_205_1994_Modified_K)
}
\description{
A simulated data for sleep study
}
\keyword{datasets}
