% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pattern_Search.R
\name{Pattern_Search}
\alias{Pattern_Search}
\title{A Function to Automatically Search for the Maximum Patterns among Strings}
\usage{
Pattern_Search(x, Direction = "Both", Match_All = TRUE, ...)
}
\arguments{
\item{x}{A vector of strings}

\item{Direction}{Patterns search from the begining of each string ("Forward"), from the end ("Backward") or "Both" ways. <<default: "Both">}

\item{Match_All}{A logical parameter indicating whether to only report patterns matched in all elements in x. If FALSE, all possible patterns will be reported. <default: TRUE>}
}
\value{
Import files into your working enviroment in R (i.e., Global Enviroment in R Studio)
}
\description{
This function will loop over strings and find the maximum patterns needed.
}
