% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update}
\alias{update}
\title{update}
\usage{
update()
}
\description{
A function to check the latest release version of RHelper package and update it.
}
\details{
This function will check the latest version of RHelper and update if needed.
}
\keyword{RHelper}
\keyword{Update,}
